# swagger.yaml
openapi: 3.0.0
info: 
  version: '1.0.0'
  title: Learning Platform
  description: this is a kind of self show case platform, there teacher can showcase there teaching style, experience, qualification and also student can evaluate and choice there comfort and prefered.
  contact: 
    name: 'Mahedi Hasan Riad'
    url: https://riad.com
    email: contact.mahedihasan@gmail.com 
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: Local host server 
    url: http://localhost:3000/api/v1
...
paths: 
...
  '/users/{id}':
    get: 
      security:
      - bearerAuth: []
      tags: [User]
      description: Find a single User
      parameters: 
        - $ref: '#/components/parameters/idPathParams'
      
      responses: 
        '200':
          description: success
          content: 
            'application/json':
              schema:
                type: object
                properties: 
                  code:
                    type: number
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: success
                  userData:
                    $ref: '#/components/schemas/userDTO'
                  link:
                    type: string
                    format: url
                    example: user/1
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    put: 
      security:
        - bearerAuth: []
      tags: [User]
      description: Update or Create a user
      parameters: 
        - $ref: '#/components/parameters/idPathParams'
      requestBody: 
        description: provide a valid info 
        content: 
          'application/json':
            schema:
              type: object
              properties: 
                name:
                  type: string
                  example: Riad
                email:
                  type: string
                  format: email
                  example: riad@gmail.com
                password:
                  type: string
                  example: riad123 
                mobile: 
                  type: number
                  format: int32
                  example: 01518949131
                avatar:
                  type: string
                  format: url
                  example: https://cloudinary.com/adiokd24e6s2d6
                cover-image:
                  type: string
                  format: url
                  example: https://cloudinary.com/adiokd24e6s2d6
                role:
                  type: string
                  enum: [Admin, teacher, coaching-admin, coaching-manager, couching-teacher, student]
                  example: student
              required: [name, email, password, mobile, avatar, role]
      responses: 
        '200':
          description: successfully updated
          content: 
            'application/json':
              schema: 
                type: object
                properties: 
                  code:
                    type: number
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: successfully updated or create 
                  info:
                    $ref: '#/components/schemas/userDTO'
                  link:
                    type: string
                    format: url
                    example: user/1
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'  
    patch: 
      security: 
        - bearerAuth: []
      tags: [User]
      description: Update a user 
      parameters: 
        - $ref: '#/components/parameters/idPathParams'
      requestBody: 
        description: provide content which one want to update 
        content: 
          'application/json':
            schema: 
              $ref: '#/components/schemas/userDTO'
      responses: 
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: object
                properties: 
                  code: 
                    type: number
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: success
                  userDate:
                    $ref: '#/components/schemas/userDTO'
                  link:
                    type: string
                    format: url
                    example: user/1    
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'       
    post: 
      security:
        - bearerAuth: []
      tags: [User]
      description: update/change password
      parameters: 
        - $ref: '#/components/parameters/idPathParams'
      requestBody: 
        description: provide currect old password to change 
        content: 
          'application/json':
            schema: 
              type: object
              properties: 
                oldPass: 
                  type: string
                  example: riad123 
                newPass:
                  type: string
                  example: riad@@@
              required: [oldPass, newPass]
      responses: 
        '200':
          description: successfully updated password 
          content: 
            'application/json':
              schema:
                type: object
                properties: 
                  code: 
                    type: number
                    format: int32
                    example: 200
                  message: 
                    type: string
                    example: successfully updated password
                  user:
                    type: object
                    properties: 
                      name: 
                        type: string
                        example: Riad
                      password:
                        type: string
                        example: riad@@@
                      link:
                        type: string
                        format: url
                        example: user/1  
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete: 
      security:
        - bearerAuth: []
      tags: [User]
      description: Delete a User
      parameters: 
        - $ref: '#/components/parameters/idPathParams'
      
      responses: 
        '204':
          description: success 
          content: 
            'application/json':
              schema: 
                type: object
                properties: 
                  name:
                    type: string
                    example: riad
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'        
...
components: 
  securitySchemes: 
    bearerAuth:
      type: http
      scheme: bearer 
      bearerFormat: JWT
  schemas: 
...
    userDTO:
      type: object
      properties: 
        name:
          type: string
          example: Riad
        email:
          type: string
          format: email
          example: riad@gmail.com
        mobile: 
          type: number
          format: int32
          example: 01518949131
        avatar:
          type: string
          format: url
          example: https://cloudinary.com/adiokd24e6s2d6
        cover-image:
          type: string
          format: url
          example: https://cloudinary.com/adiokd24e6s2d6
        role:
          type: string
          enum: [Admin, teacher, coaching-admin, coaching-manager, couching-teacher, student]
          example: student
...
  responses: 
...
    '400':
      description: Bad Request
      content: 
        'application/json':
          schema: 
            type: object
            properties: 
              code:
                type: number
                format: int32
                example: 400
              message:
                type: string
                example: Bad Request
              data:
                type: object
                properties: 
                  message:
                    type: string
                    example: error
                  field:
                    type: string
                    example: field name
                  in:
                    type: string
                    enum: [Query, Body]
                    example: Query 
    '401':
      description: unauthorized
      content: 
        'application/json':
          schema: 
            type: object
            properties: 
              code: 
                type: number
                format: int32
                example: 401
              message:
                type: string
                example: unauthorized
    '404':
      description: not found
      content: 
        'application/json':
          schema: 
            type: object
            properties: 
              code:
                type: number
                format: int32
                example: 404
              message:
                type: string
                example: Content Not found
    '500':
      description: Server Error 
      content: 
        'application/json':
          schema: 
            type: object
            properties: 
              code: 
                type: number
                format: int32
                example: 500
              message:
                type: string
                example: Server Error
  parameters: 
    'idPathParams':
      in: path
      name: id 
      description: user / article id 
      required: true
      schema: 
        type: number
        format: int32
        example: 1
...